// packages/wizzi.studio/src/features/dev/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi.plugins\packages\wizzi.plugin.ts\lib\artifacts\ts\module\gen\main.js
    package: wizzi.plugin.ts@
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.apps\packages\wizzi.studio\.wizzi-override\src\features\dev\types.ts.ittf
    utc time: Sat, 17 Feb 2024 04:55:15 GMT
*/
export type ScanDevFoldersOptions = { 
    hasPlugins?: boolean;
    hasPluginContents?: boolean;
    hasMetaPlugins?: boolean;
    hasMetaPluginContents?: boolean;
};